Make fly io review apps idempotent on pg trb 24 (#5)
This PR does three different things and it's a bit hard to see in the
movement of the code.

First, the cluster creation will only be done once just after the
database creation and since it's now under the "if db not exists", it
will only run when it does not exists.

Second, I removed the `|| true` of the DB creation: if it does not
create the DB when we want it to be, it should be a fail.

Third, I moved the `flyctl attach` command to be run everytime. This is
because shit happens in Fly and sometimes the DB is not attached
properly and just running it again could fix it.