Strategy immediate and avoid copy config (#7)
This PR changes two things in our deploy method:
- set the strategy to "immediate"
- removes the `--copy-config`

Firstly, the change to `--strategy immediate` is because of a Fly bug.
Ideally, we would set secrets before the first deploy and we could use
the rolling strategy. But since we sometimes have secrets not set when
we set them before the first deploy, we need to set them after.
Therefore, we have a first deploy without all needed secrets and errors
when we try to start the instances.

If we're in rolling strategy, this will stop the deployment and secrets
won't be set.
So for now, we need to keep the immediate strategy.

Secondly, removing the `--copy-config` will avoid having Fly creates a
`fly.toml` with an app name and trigger interactive prompts because the
app name is not the same for the postgres app.